+++ lib/memory.c
@@ -1,5 +1,6 @@
 #include "../include/memory.h"
 #include <stdint.h>
+#include <stddef.h>
 
 void * memcpy(void * dest, void * src, size_t len) {
     uint8_t * d = dest;
@@ -10,7 +11,7 @@ void * memcpy(void * dest, void * src, size_t len) {
 }
 
 
-volatile void * memcpy_vd(volatile void * dest, void * src, unsigned len) {
+volatile void * memcpy_vd(volatile void * dest, void * src, size_t len) {
     volatile uint8_t * d = dest;
     while (len--)
         *(d++) = *((uint8_t* ) src++);
@@ -18,7 +19,7 @@ volatile void * memcpy_vd(volatile void * dest, void * src, unsigned len) {
     return dest;
 }
 
-void * memcpy_vs(void * dest, volatile void * src, unsigned len) {
+void * memcpy_vs(void * dest, volatile void * src, size_t len) {
     uint8_t * d = dest;
     while (len--)
         *(d++) = *((volatile uint8_t *) src++);
@@ -26,7 +27,7 @@ void * memcpy_vs(void * dest, volatile void * src, unsigned len) {
     return dest;
 }
 
-volatile void * memcpy_vds(volatile void * dest, volatile void * src, unsigned len) {
+volatile void * memcpy_vds(volatile void * dest, volatile void * src, size_t len) {
     volatile uint8_t * d = dest;
     while (len--)
         *(d++) = *((volatile uint8_t *) src++);
@@ -34,7 +35,6 @@ volatile void * memcpy_vds(volatile void * dest, volatile void * src, unsigned l
     return dest;
 }
 
-    #include "../src/wd_debug.h"
 void * memset(void * dest, unsigned char val, size_t len) {
     uint8_t * d = dest;
     while (len--)
